added timings

+++ dA_kernel.cu
@@ -40,6 +40,7 @@ __global__ void dA_train_kernel(dA model, float *X_d, double learning_rate, doub
     L_hbias[tid] = 0.0;
   }
   __syncthreads();
+
   // tilde_x: da_get_corrupted_input()
   __shared__ float tilde_x[N_FEATS];
   if (X_d[start + bid*N_FEATS + tid] == 0) {
@@ -57,11 +58,9 @@ __global__ void dA_train_kernel(dA model, float *X_d, double learning_rate, doub
   __syncthreads();
   for (int h=0; h < N_HIDDEN; ++h) {
     atomicAdd(&y[h], model.W_flat[h*N_FEATS + tid] * X_d[start + bid*N_FEATS + tid]);
-    //y[h] += model.W_flat[h*N_FEATS + tid] * X_d[start + bid*N_FEATS + tid];
   }
   __syncthreads();
   if (tid < N_HIDDEN) {
-    //atomicAdd(&y[tid], model.hbias[tid]);
     y[tid] += model.hbias[tid];
     y[tid] = sigmoid_kernel(y[tid]);
   }
@@ -71,11 +70,9 @@ __global__ void dA_train_kernel(dA model, float *X_d, double learning_rate, doub
   __shared__ double z[N_FEATS];
   z[tid] = 0.0;
   for (int h=0; h < N_HIDDEN; ++h) {
-    //atomicAdd(&z[tid], model.W_flat[h*N_FEATS + tid] * y[h]);
     z[tid] += model.W_flat[h*N_FEATS + tid] * y[h];
   }
   __syncthreads();
-  //atomicAdd(&z[tid], model.vbias[tid]);
   z[tid] += model.vbias[tid];
   z[tid] = sigmoid_kernel(z[tid]);
   __syncthreads();
@@ -83,27 +80,24 @@ __global__ void dA_train_kernel(dA model, float *X_d, double learning_rate, doub
 
   // update vbias, each thread grabs a column of X_d, each block works on N_OBS / BATCH_SIZE rows
   L_vbias[tid] = X_d[start + bid * N_FEATS + tid] - z[tid];
-  //atomicAdd(&model.vbias[tid], learning_rate * L_vbias[tid] / model.N);
   model.vbias[tid] += learning_rate * L_vbias[tid] / model.N;
   __syncthreads();
+
 
   // update hbias
   for (int h=0; h < N_HIDDEN; ++h) {
     atomicAdd(&L_hbias[h], model.W_flat[h*N_FEATS + tid] * L_vbias[tid]);
-    //L_hbias[h] += model.W_flat[h*N_FEATS + tid] * L_vbias[tid];
   }
   if (tid < N_HIDDEN) {
-    //atomicMultiply(&L_hbias[tid], y[tid] * (1.0 - y[tid]));
-    //atomicAdd(&model.hbias[tid], learning_rate * L_hbias[tid] / model.N);
     L_hbias[tid] *= y[tid] * (1.0 - y[tid]);
     model.hbias[tid] += learning_rate * L_hbias[tid] / model.N;
   }
   __syncthreads();  
 
+
   // Update weights
   for (int h=0; h < N_HIDDEN; ++h) {
-    atomicAdd(&model.W_flat[h*N_FEATS + tid], learning_rate * (L_hbias[h] * tilde_x[tid] + L_vbias[tid] * y[h]) / (double)model.N / (double)BATCH_SIZE);
-    //model.W_flat[h * N_FEATS + tid] += learning_rate * (L_hbias[h] * tilde_x[tid] + L_vbias[tid] * y[h]) / model.N;
+    atomicAdd(&model.W_flat[h*N_FEATS + tid], learning_rate * (L_hbias[h] * tilde_x[tid] + L_vbias[tid] * y[h]) / model.N / BATCH_SIZE);
   }
 
 }
